\name{GenCAT-package}
\alias{GenCAT-package}
\alias{GenCAT-package}
\docType{package}
\title{
\packageTitle{GenCAT}
}
\description{
\packageDescription{GenCAT}
}
\details{


\packageDESCRIPTION{GenCAT}

\packageIndices{GenCAT}
}
\author{
\packageAuthor{GenCAT}

Maintainer: \packageMaintainer{GenCAT}
}
\references{
Qian J, Nunez S, Reed E, Reilly MP, Foulkes AS (2016) <DOI:10.1371/journal.pone.0148218> A Simple Test of Class-Level Genetic Association Can Reveal Novel Cardiometabolic Trait Loci. PLoS ONE 11(2): e0148218.

Rosenbloom, K. R., Armstrong, J., Barber, G. P., Casper, J., Clawson, H., Diekhans, M., Dreszer, T. R., Fujita, P. A., Guruvadoo, L., Haeussler, M., Harte, R. A., Heitner, S., Hickey, G., Hinrichs, A. S., Hubley, R., Karolchik, D., Learned, K., Lee, B. T., Li, C. H., Miga, K.H. , Nguyen, N., Paten, B., Raney, B.J., Smit, A. F., Speir, M. L., Zweig, A. S., Haussler, D., Kuhn, R. M., Kent, W.J. (2015) <DOI:10.1093/nar/gku1177> The UCSC Genome Browser database: 2015 update. Nucleic Acids Res. Jan;43(Database issue):D670-81.

Schunkert, H., Konig, I. R., Kathiresan, S., Reilly, M. P. and et. al. (2011) <DOI:10.1038/ng.784> Large-scale association analysis identifies 13 new susceptibility loci for coronary artery disease. Nat. Genet. 43 333-338.

The 1000 Genomes Project Consortium. (2012) <DOI:10.1038/nature11632> An integrated map of genetic variation from 1,092 human genomes. Nature, 491(7422), 56-65.

}

\keyword{ package }

