\name{geno}
\alias{geno}
\docType{data}
\title{Genotype Data
}
\description{
This is genotype data as read in by read.plink (snpStats).  It contains genotypes for 99 individuals from the CEU population (Utah residents with northern and western European ancestry).  It only includes chromosomes 13-15. 
}
\usage{data("geno")}
\format{
  The format is:
List of 3
 $ genotypes:Formal class 'SnpMatrix' [package "snpStats"] with 1 slot
  .. ..@ .Data: raw [1:99, 1:84195] 01 03 02 03 ...
  .. .. ..- attr(*, "dimnames")=List of 2
  .. .. .. ..$ : chr [1:99] "CEU_1" "CEU_2" "CEU_3" "CEU_4" ...
  .. .. .. ..$ : chr [1:84195] "rs624673" "rs9511877" "rs638773" "rs9511880" ...
 $ fam      :'data.frame':	99 obs. of  6 variables:
  ..$ pedigree: chr [1:99] "CEU_1" "CEU_2" "CEU_3" "CEU_4" ...
  ..$ member  : int [1:99] 1 1 1 1 1 1 1 1 1 1 ...
  ..$ father  : logi [1:99] NA NA NA NA NA NA ...
  ..$ mother  : logi [1:99] NA NA NA NA NA NA ...
  ..$ sex     : logi [1:99] NA NA NA NA NA NA ...
  ..$ affected: logi [1:99] NA NA NA NA NA NA ...
 $ map      :'data.frame':	84195 obs. of  6 variables:
  ..$ chromosome: int [1:84195] 13 13 13 13 13 13 13 13 13 13 ...
  ..$ snp.name  : chr [1:84195] "rs624673" "rs9511877" "rs638773" "rs9511880" ...
  ..$ cM        : logi [1:84195] NA NA NA NA NA NA ...
  ..$ position  : num [1:84195] 19743996 19744070 19744848 19745096 19745251 ...
  ..$ allele.1  : chr [1:84195] "G" "A" "A" "G" ...
  ..$ allele.2  : chr [1:84195] "A" "G" "G" "T" ...
}
\details{
This data was filtered for: minor allele frequency < 0.01, call rate  < 0.95, and Hardy-Weinberg Equilibrium p < 10^-6.
}
\source{
ftp://ftp-trace.ncbi.nih.gov/1000genomes/ftp/release/20130502/
}
\references{
The 1000 Genomes Project Consortium. (2012). An integrated map of genetic variation from 1,092 human genomes. Nature, 491(7422), 56-65.
}
\examples{
data("geno")
}
\keyword{datasets}
