\name{GenCAT-package}
\alias{GenCAT-package}
\alias{GenCAT-package}
\docType{package}
\title{
\packageTitle{GenCAT}
}
\description{
\packageDescription{GenCAT}
}
\details{


\packageDESCRIPTION{GenCAT}

\packageIndices{GenCAT}
}
\author{
\packageAuthor{GenCAT}

Maintainer: \packageMaintainer{GenCAT}
}
\references{
Qian, J., Reed, E., Nunez, S., Qu, L., Reilly, M., Foulkes, A. (2016) Testing class-level genetic associations using single-element summary statistics. (Under review).

Rosenbloom, K. R., Armstrong, J., Barber, G. P., Casper, J., Clawson, H., Diekhans, M., Dreszer, T. R., Fujita, P. A., Guruvadoo, L., Haeussler, M., Harte, R. A., Heitner, S., Hickey, G., Hinrichs, A. S., Hubley, R., Karolchik, D., Learned, K., Lee, B. T., Li, C. H., Miga, K.H. , Nguyen, N., Paten, B., Raney, B.J., Smit, A. F., Speir, M. L., Zweig, A. S., Haussler, D., Kuhn, R. M., Kent, W.J. The UCSC Genome Browser database: 2015 update. Nucleic Acids Res. 2015 Jan;43(Database issue):D670-81. PMID: 25428374; PMC: PMC4383971 

Schunkert, H., Konig, I. R., Kathiresan, S., Reilly, M. P. and et. al. (2011). Large-scale association analysis identifies 13 new susceptibility loci for coronary artery disease. Nat. Genet. 43 333-338.

The 1000 Genomes Project Consortium. (2012). An integrated map of genetic variation from 1,092 human genomes. Nature, 491(7422), 56-65.

}

\keyword{ package }

